+++ stories/ElementsStories/Radio.stories.tsx
@@ -1,4 +1,4 @@
-import { ArgsTable, Story, Title } from "@storybook/blocks";
+import { Story } from "@storybook/blocks";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Radio } from "@elements/radio";
@@ -28,8 +28,9 @@ export const Default: Story = {
         <div className="hawa-flex hawa-flex-col hawa-gap-2">
           <Radio
             onChangeTab={() => console.log()}
-            defaultValue={"option2"}
-            design={"default"}
+            name="default"
+            design="default"
+            defaultValue="option2"
             labelProps={{ htmlFor: "horizontal" }}
             label="Horizontal orientation"
             helperText="Please pick an option"
@@ -39,14 +40,16 @@ export const Default: Story = {
               { label: `Option 2`, value: `option2`, disabled: true },
               { label: `Option 3`, value: `option3` }
             ]}
+            {...args}
           />
         </div>
         <div className="hawa-flex hawa-flex-col hawa-gap-2">
           <Radio
             direction={direction}
             onChangeTab={() => console.log()}
-            defaultValue={"option6"}
-            design={"default"}
+            name="default"
+            design="default"
+            defaultValue="option6"
             orientation="vertical"
             labelProps={{ htmlFor: "horizontal" }}
             label="Vertical orientation"
@@ -56,10 +59,14 @@ export const Default: Story = {
               { label: `Option 5`, value: `option5`, disabled: true },
               { label: `Option 6`, value: `option6` }
             ]}
+            {...args}
           />
         </div>
       </div>
     );
+  },
+  argTypes: {
+    onChange: { action: "onChange" }
   }
 };
 
@@ -67,7 +74,7 @@ export const Tabs: Story = {
   parameters: {
     layout: "padded"
   },
-  render: () => (
+  render: (args: any) => (
     <div className="hawa-relative hawa-flex hawa-flex-row hawa-gap-4 hawa-space-x-4">
       <div className="hawa-max-w-xs">
         <div className="hawa-pb-4 hawa-text-center hawa-text-2xl hawa-font-bold">
@@ -81,15 +88,15 @@ export const Tabs: Story = {
             </h1>
             <Radio
               size="xs"
-              onChangeTab={() => console.log()}
-              defaultValue={"option2"}
-              design={"tabs"}
-              // {...args}
+              name="tabs"
+              design="tabs"
+              defaultValue="option2"
               options={[
                 { label: `Option 1`, value: `option1` },
                 { label: `Option 2`, value: `option2`, disabled: true },
                 { label: `Option 3`, value: `option3` }
               ]}
+              {...args}
             />
           </div>
           <div className="hawa-flex hawa-flex-col hawa-gap-2">
@@ -98,15 +105,16 @@ export const Tabs: Story = {
             </h1>
             <Radio
               size="xs"
-              onChangeTab={() => console.log()}
-              defaultValue={"option6"}
-              design={"tabs"}
+              name="tabs"
+              design="tabs"
+              defaultValue="option6"
               orientation="vertical"
               options={[
                 { label: `Option 4`, value: `option4` },
                 { label: `Option 5`, value: `option5`, disabled: true },
                 { label: `Option 6`, value: `option6` }
               ]}
+              {...args}
             />
           </div>
         </div>
@@ -125,15 +133,15 @@ export const Tabs: Story = {
             </h1>
             <Radio
               size="sm"
-              onChangeTab={() => console.log()}
-              defaultValue={"option2"}
-              design={"tabs"}
-              // {...args}
+              name="tabs"
+              design="tabs"
+              defaultValue="option2"
               options={[
                 { label: `Option 1`, value: `option1` },
                 { label: `Option 2`, value: `option2`, disabled: true },
                 { label: `Option 3`, value: `option3` }
               ]}
+              {...args}
             />
           </div>
           <div className="hawa-flex hawa-flex-col hawa-gap-2">
@@ -142,15 +150,16 @@ export const Tabs: Story = {
             </h1>
             <Radio
               size="sm"
-              onChangeTab={() => console.log()}
-              defaultValue={"option6"}
-              design={"tabs"}
+              name="tabs"
+              design="tabs"
+              defaultValue="option6"
               orientation="vertical"
               options={[
                 { label: `Option 4`, value: `option4` },
                 { label: `Option 5`, value: `option5`, disabled: true },
                 { label: `Option 6`, value: `option6` }
               ]}
+              {...args}
             />
           </div>
         </div>
@@ -168,15 +177,15 @@ export const Tabs: Story = {
               Orientation: <strong>Horizontal</strong>
             </h1>
             <Radio
-              onChangeTab={() => console.log()}
-              defaultValue={"option2"}
-              design={"tabs"}
-              // {...args}
+              name="tabs"
+              design="tabs"
+              defaultValue="option2"
               options={[
                 { label: `Option 1`, value: `option1` },
                 { label: `Option 2`, value: `option2`, disabled: true },
                 { label: `Option 3`, value: `option3` }
               ]}
+              {...args}
             />
           </div>
           <div className="hawa-flex hawa-flex-col hawa-gap-2">
@@ -184,15 +193,16 @@ export const Tabs: Story = {
               Orientation: <strong>Vertical</strong>
             </h1>
             <Radio
-              onChangeTab={() => console.log()}
-              defaultValue={"option6"}
-              design={"tabs"}
+              name="tabs"
+              design="tabs"
+              defaultValue="option6"
               orientation="vertical"
               options={[
                 { label: `Option 4`, value: `option4` },
                 { label: `Option 5`, value: `option5`, disabled: true },
                 { label: `Option 6`, value: `option6` }
               ]}
+              {...args}
             />
           </div>
         </div>
@@ -211,15 +221,15 @@ export const Tabs: Story = {
             </h1>
             <Radio
               size="lg"
-              onChangeTab={() => console.log()}
-              defaultValue={"option2"}
-              design={"tabs"}
-              // {...args}
+              name="tabs"
+              design="tabs"
+              defaultValue="option2"
               options={[
                 { label: `Option 1`, value: `option1` },
                 { label: `Option 2`, value: `option2`, disabled: true },
                 { label: `Option 3`, value: `option3` }
               ]}
+              {...args}
             />
           </div>
           <div className="hawa-flex hawa-flex-col hawa-gap-2">
@@ -228,21 +238,25 @@ export const Tabs: Story = {
             </h1>
             <Radio
               size="lg"
-              onChangeTab={() => console.log()}
-              defaultValue={"option6"}
-              design={"tabs"}
+              name="tabs"
+              design="tabs"
+              defaultValue="option6"
               orientation="vertical"
               options={[
                 { label: `Option 4`, value: `option4` },
                 { label: `Option 5`, value: `option5`, disabled: true },
                 { label: `Option 6`, value: `option6` }
               ]}
+              {...args}
             />
           </div>
         </div>
       </div>
     </div>
-  )
+  ),
+  argTypes: {
+    onChange: { action: "onChange" }
+  }
 };
 export const Cards: Story = {
   render: (args: any, globals: any) => {
@@ -260,9 +274,9 @@ export const Cards: Story = {
             Orientation: <strong>Horizontal</strong>
           </h1>
           <Radio
-            design={"cards"}
-            defaultValue={"option2"}
-            onChangeTab={() => console.log()}
+            name="cards"
+            design="cards"
+            defaultValue="option2"
             options={[
               {
                 label: `Option 1`,
@@ -281,6 +295,7 @@ export const Cards: Story = {
                 value: `option3`
               }
             ]}
+            {...args}
           />
         </div>
         <div className="hawa-flex hawa-flex-col hawa-gap-2">
@@ -288,10 +303,10 @@ export const Cards: Story = {
             Orientation: <strong>Vertical</strong>
           </h1>
           <Radio
-            design={"cards"}
+            name="cards"
+            design="cards"
             orientation="vertical"
-            defaultValue={"option6"}
-            onChangeTab={() => console.log()}
+            defaultValue="option6"
             options={[
               {
                 label: `Option 4`,
@@ -310,10 +325,14 @@ export const Cards: Story = {
                 value: `option6`
               }
             ]}
+            {...args}
           />
         </div>
       </div>
     );
+  },
+  argTypes: {
+    onChange: { action: "onChange" }
   }
 };
 export const Bordered: Story = {
@@ -332,9 +351,9 @@ export const Bordered: Story = {
             Orientation: <strong>Horizontal</strong>
           </h1>
           <Radio
-            design={"bordered"}
-            defaultValue={"option2"}
-            onChangeTab={() => console.log()}
+            name="bordered"
+            design="bordered"
+            defaultValue="option2"
             direction={direction}
             options={[
               {
@@ -354,6 +373,7 @@ export const Bordered: Story = {
                 value: `option3`
               }
             ]}
+            {...args}
           />
         </div>
         <div className="hawa-flex hawa-flex-col hawa-gap-2">
@@ -361,10 +381,10 @@ export const Bordered: Story = {
             Orientation: <strong>Vertical</strong>
           </h1>
           <Radio
-            design={"bordered"}
+            name="bordered"
+            design="bordered"
             orientation="vertical"
-            defaultValue={"option6"}
-            onChangeTab={() => console.log()}
+            defaultValue="option6"
             direction={direction}
             options={[
               {
@@ -384,9 +404,13 @@ export const Bordered: Story = {
                 value: `option6`
               }
             ]}
+            {...args}
           />
         </div>
       </div>
     );
+  },
+  argTypes: {
+    onChange: { action: "onChange" }
   }
 };